refactor(cli): extract link to action
